Fix install process in README.md
Fix replaces unreliable though common methodology of piping ps aux output through grep to locate pg_hba.conf. This method will not work on all systems, as it relies on a tagentially related query to locate the file. The only reliable method for locating pg_hba.conf is to query directly via the postgres shell using the SHOW hba_file; command, which is specifically designed for this purpose.