Handle charge dates not migrated as well as unpaid & past_due
We want to continue to charge pre-migration folks the day they signed
up. Because of that, setting `nextChargeDate' and `nextPeriodStart' to
the first day of the next month should only happen when subscriptions
are being created.

The `charge_subscriptions' script will only increment the charge dates
by a month or year, it won't find the first day of that month anymore.

The unpaid & past_due subscriptions are going to be charged as soon as
the migration runs, however, their `nextPeriodStart' doesn't start
until the end of the current billing period from stripe.