charge_subscriptions: Ensure subscription is always updated
There was an error sending a confirmation email after successfully
charging a subscription. That error caused the whole script to stop
and no subscriptions on that batch were updated. I manually checked
which subscriptions were charged on Stripe and updated the database
accordingly.

From now one this kind of error will be logged but it won't interrupt
the script anymore.