Fix saving transaction.data field & use sequelize transactions
The data field wasn't being saved because sequelize has a little bit
of a problem to detect changes into JSON fields so it was needed to
explicitly call the method `changed` every time the field was updated.

Also, it's a lil faster to bundle a batch within a single database
transaction.