infra: Add environment aware scripts for restoring database dumps.
Most of the script `db_setup.js` was refactored into a small library
called `lib/db.js` and most of its code is now shared with the new
script introduced in this commit: `db_restore.js` which does the same
thing as its bash alternative. The only difference is that it's aware
of the environment and it has access to the configuration system.