db_restore.js: Don't force restoring by default
The db_restore.js script now runs on the `predev` role and it kinda
breaks the whole setup if the database is being used (by a database
visualization tool for example) because it tries to recreate the
database.

This patch adds the `-f` flag to the `db_restore.js` script and makes
it less intrusive by default, not trying to recreate anything if data
exists.