chore(stores): Improve API for creating hosts & collectives
The creation of collectives & hosts at the same time is indeed a
wanted feature. But creating a host and then creating different
collectives is also very useful too. To make both ways possible, a new
function was added to the test stores: `newHost`.

Also, the function `hostAndCollective` was renamed to
`newCollectiveWithHost` and the function `collectiveWithHost` was
renamed to `newCollectiveInHost` to increase the consistency of the
function names.