Introduce a "Reply to" email for host accounts (#8006)
* CC admins when approving collectives

When I am approving or rejecting a collective to join my fiscal host, I would like to get an email in Freshdesk from the message which I send using the platform. I believe this enables that functionality.

I am unsure about the replyTo being the no-reply email address, and whether or not HostCollectiveID is the right email. Ideally, this would be my email, as the host admin making the call. Is that possible?

* feat(email): add reply to address for fiscal hosts

Co-authored-by: Benjamin Piouffle <benjamin@opencollective.com>