Expense tag stats (#9426)
Project: https://github.com/opencollective/opencollective/issues/7033

### Description
Adding a `expenseTagStats` root query to support an autocompleting fuzzy search filter on expense tags in the expenses dashboard, for both hosts and regular accounts.

It's returning results from a materialized view `ExpenseTagStats`, that has tag counts grouped on either HostCollectiveId and CollectiveId.

While the `account.expensesTags` field could potentially be extended to include a `tagSearchTerm`, it might not be ideal to do a fairly expensive query in an autocompleting search. It also lacks support for the host use case.